revert clmm/dataops/data_operations.py -> clmm/dataops/__init__.py
